fixed: Запись в архивы значений типа float

+++ src/user/log/log_event.cpp
@@ -12,7 +12,7 @@ LogEvent::~LogEvent()
 }
 
 void LogEvent::add(uint8_t code, uint8_t type, uint16_t id,
-                   uint32_t oldValue, uint32_t newValue)
+                   float oldValue, float newValue)
 {
   time_t time = getTime();
 
@@ -21,8 +21,8 @@ void LogEvent::add(uint8_t code, uint8_t type, uint16_t id,
   *(uint8_t*)(buffer+8) = code;
   *(uint8_t*)(buffer+9) = type;
   *(uint16_t*)(buffer+10) = id;
-  *(uint32_t*)(buffer+12) = oldValue;
-  *(uint32_t*)(buffer+16) = newValue;
+  *(float*)(buffer+12) = oldValue;
+  *(float*)(buffer+16) = newValue;
   write(buffer, 20);
 }
