+++ src/user/log/log_tms.cpp
@@ -42,9 +42,9 @@ void LogTms::add()
 
   *(uint32_t*)(buffer) = ++id_;
   *(uint32_t*)(buffer+4) = time;
-  *(uint32_t*)(buffer+8) = code;
-  *(uint32_t*)(buffer+9) = KSU.getValue(TMS_PRESSURE_INTAKE);
-  *(uint32_t*)(buffer+13) = KSU.getValue(TMS_TEMPERATURE_WINDING);
+  *(uint8_t*)(buffer+8) = code;
+  *(float*)(buffer+9) = KSU.getValue(TMS_PRESSURE_INTAKE);
+  *(float*)(buffer+13) = KSU.getValue(TMS_TEMPERATURE_WINDING);
 
   write(buffer, 17);
 }
